TST/REF: collect tests by method from generic (#37421)
