+++ pandas/tests/generic/test_frame.py
@@ -15,13 +15,6 @@ class TestDataFrame(Generic):
     _typ = DataFrame
     _comparator = lambda self, x, y: tm.assert_frame_equal(x, y)
 
-    def test_rename_mi(self):
-        df = DataFrame(
-            [11, 21, 31],
-            index=MultiIndex.from_tuples([("A", x) for x in ["a", "B", "c"]]),
-        )
-        df.rename(str.lower)
-
     @pytest.mark.parametrize("func", ["_set_axis_name", "rename_axis"])
     def test_set_axis_name(self, func):
         df = DataFrame([[1, 2], [3, 4]])
@@ -76,8 +69,7 @@ class TestDataFrame(Generic):
         expected = DataFrame(index=[0, 1, 2], dtype=object)
         self._compare(result, expected)
 
-    def test_metadata_propagation_indiv(self):
-
+    def test_metadata_propagation_indiv_groupby(self):
         # groupby
         df = DataFrame(
             {
@@ -90,6 +82,7 @@ class TestDataFrame(Generic):
         result = df.groupby("A").sum()
         self.check_metadata(df, result)
 
+    def test_metadata_propagation_indiv_resample(self):
         # resample
         df = DataFrame(
             np.random.randn(1000, 2),
@@ -98,6 +91,7 @@ class TestDataFrame(Generic):
         result = df.resample("1T")
         self.check_metadata(df, result)
 
+    def test_metadata_propagation_indiv(self):
         # merging with override
         # GH 6923
         _metadata = DataFrame._metadata
