+++ pandas/tests/generic/test_series.py
@@ -14,13 +14,6 @@ class TestSeries(Generic):
     _typ = Series
     _comparator = lambda self, x, y: tm.assert_series_equal(x, y)
 
-    def test_rename_mi(self):
-        s = Series(
-            [11, 21, 31],
-            index=MultiIndex.from_tuples([("A", x) for x in ["a", "B", "c"]]),
-        )
-        s.rename(str.lower)
-
     @pytest.mark.parametrize("func", ["rename_axis", "_set_axis_name"])
     def test_set_axis_name_mi(self, func):
         s = Series(
@@ -104,17 +97,7 @@ class TestSeries(Generic):
             with pytest.raises(ValueError, match=msg):
                 s.bool()
 
-    def test_metadata_propagation_indiv(self):
-        # check that the metadata matches up on the resulting ops
-
-        o = Series(range(3), range(3))
-        o.name = "foo"
-        o2 = Series(range(3), range(3))
-        o2.name = "bar"
-
-        result = o.T
-        self.check_metadata(o, result)
-
+    def test_metadata_propagation_indiv_resample(self):
         # resample
         ts = Series(
             np.random.rand(1000),
@@ -129,6 +112,17 @@ class TestSeries(Generic):
 
         result = ts.resample("1T").apply(lambda x: x.sum())
         self.check_metadata(ts, result)
+
+    def test_metadata_propagation_indiv(self):
+        # check that the metadata matches up on the resulting ops
+
+        o = Series(range(3), range(3))
+        o.name = "foo"
+        o2 = Series(range(3), range(3))
+        o2.name = "bar"
+
+        result = o.T
+        self.check_metadata(o, result)
 
         _metadata = Series._metadata
         _finalize = Series.__finalize__
@@ -157,25 +151,3 @@ class TestSeries(Generic):
         # reset
         Series._metadata = _metadata
         Series.__finalize__ = _finalize  # FIXME: use monkeypatch
-
-
-class TestSeries2:
-    # Separating off because it doesnt rely on parent class
-    @pytest.mark.parametrize(
-        "s",
-        [
-            Series([np.arange(5)]),
-            pd.date_range("1/1/2011", periods=24, freq="H"),
-            Series(range(5), index=pd.date_range("2017", periods=5)),
-        ],
-    )
-    @pytest.mark.parametrize("shift_size", [0, 1, 2])
-    def test_shift_always_copy(self, s, shift_size):
-        # GH22397
-        assert s.shift(shift_size) is not s
-
-    @pytest.mark.parametrize("move_by_freq", [pd.Timedelta("1D"), pd.Timedelta("1M")])
-    def test_datetime_shift_always_copy(self, move_by_freq):
-        # GH22397
-        s = Series(range(5), index=pd.date_range("2017", periods=5))
-        assert s.shift(freq=move_by_freq) is not s
