+++ pandas/tests/series/methods/test_shift.py
@@ -19,6 +19,25 @@ from pandas.tseries.offsets import BDay
 
 
 class TestShift:
+    @pytest.mark.parametrize(
+        "ser",
+        [
+            Series([np.arange(5)]),
+            date_range("1/1/2011", periods=24, freq="H"),
+            Series(range(5), index=date_range("2017", periods=5)),
+        ],
+    )
+    @pytest.mark.parametrize("shift_size", [0, 1, 2])
+    def test_shift_always_copy(self, ser, shift_size):
+        # GH22397
+        assert ser.shift(shift_size) is not ser
+
+    @pytest.mark.parametrize("move_by_freq", [pd.Timedelta("1D"), pd.Timedelta("1M")])
+    def test_datetime_shift_always_copy(self, move_by_freq):
+        # GH#22397
+        ser = Series(range(5), index=date_range("2017", periods=5))
+        assert ser.shift(freq=move_by_freq) is not ser
+
     def test_shift(self, datetime_series):
         shifted = datetime_series.shift(1)
         unshifted = shifted.shift(-1)
